Configurable endpoint
Per discussion in https://github.com/transloadit/uppy/pull/663.

This extends the current approach with an `AWS_ENDPOINT` environment
variable that you can use to configure the S3 endpoint. The AWS SDK
helpfully supports endpoints in this template format so regions can
still be specified as a separate variable:

    https://{region}.digitaloceanspaces.com

As mentioned in transloadit/uppy#663 we may need something more flexible
in the future so you don't have to run a separate server for every
bucket or w/e, but this is a low-effort way to support S3 alternatives
for now, which I think is worth something.